// src/app/Models/Product.ts
export class Product {
    id: number;
    name: string;
    description: string;
    brand: string;
    gender: string;
    category: string;
    size: string;
    color: number;
    price: number;
    isInInventory: boolean;
    discounted: number;
    itemsLeft: number;
    imageUrl: string;
    slug: string
}